Move top-level native program tests to their respective crates
