+++ programs/tests/programs.rs
@@ -1,99 +1,3 @@
-use solana_runtime::bank::Bank;
-use solana_runtime::bank::BankError;
-use solana_sdk::genesis_block::GenesisBlock;
-use solana_sdk::loader_transaction::LoaderTransaction;
-use solana_sdk::native_loader;
-use solana_sdk::native_program::ProgramError;
-use solana_sdk::pubkey::Pubkey;
-use solana_sdk::signature::{Keypair, KeypairUtil};
-use solana_sdk::system_transaction::SystemTransaction;
-use solana_sdk::transaction::Transaction;
-
-fn load_program(bank: &Bank, from: &Keypair, loader_id: Pubkey, program: Vec<u8>) -> Pubkey {
-    let program_account = Keypair::new();
-
-    let tx = SystemTransaction::new_program_account(
-        from,
-        program_account.pubkey(),
-        bank.last_blockhash(),
-        1,
-        program.len() as u64,
-        loader_id,
-        0,
-    );
-    bank.process_transaction(&tx).unwrap();
-    assert_eq!(bank.get_signature_status(&tx.signatures[0]), Some(Ok(())));
-
-    let chunk_size = 256; // Size of chunk just needs to fit into tx
-    let mut offset = 0;
-    for chunk in program.chunks(chunk_size) {
-        let tx = LoaderTransaction::new_write(
-            &program_account,
-            loader_id,
-            offset,
-            chunk.to_vec(),
-            bank.last_blockhash(),
-            0,
-        );
-        bank.process_transaction(&tx).unwrap();
-        offset += chunk_size as u32;
-    }
-
-    let tx = LoaderTransaction::new_finalize(&program_account, loader_id, bank.last_blockhash(), 0);
-    bank.process_transaction(&tx).unwrap();
-    assert_eq!(bank.get_signature_status(&tx.signatures[0]), Some(Ok(())));
-
-    program_account.pubkey()
-}
-
-#[test]
-fn test_program_native_noop() {
-    solana_logger::setup();
-
-    let (genesis_block, mint_keypair) = GenesisBlock::new(50);
-    let bank = Bank::new(&genesis_block);
-
-    let program = "noop".as_bytes().to_vec();
-    let program_id = load_program(&bank, &mint_keypair, native_loader::id(), program);
-
-    // Call user program
-    let tx = Transaction::new(
-        &mint_keypair,
-        &[],
-        program_id,
-        &1u8,
-        bank.last_blockhash(),
-        0,
-    );
-    bank.process_transaction(&tx).unwrap();
-    assert_eq!(bank.get_signature_status(&tx.signatures[0]), Some(Ok(())));
-}
-
-#[test]
-fn test_program_native_failure() {
-    solana_logger::setup();
-
-    let (genesis_block, mint_keypair) = GenesisBlock::new(50);
-    let bank = Bank::new(&genesis_block);
-
-    let program = "failure".as_bytes().to_vec();
-    let program_id = load_program(&bank, &mint_keypair, native_loader::id(), program);
-
-    // Call user program
-    let tx = Transaction::new(
-        &mint_keypair,
-        &[],
-        program_id,
-        &1u8,
-        bank.last_blockhash(),
-        0,
-    );
-    assert_eq!(
-        bank.process_transaction(&tx),
-        Err(BankError::ProgramError(0, ProgramError::GenericError))
-    );
-}
-
 #[cfg(feature = "bpf_c")]
 use solana_sdk::bpf_loader;
 #[cfg(any(feature = "bpf_c", feature = "bpf_rust"))]
