+++ runtime/src/lib.rs
@@ -3,6 +3,7 @@ pub mod appendvec;
 pub mod bank;
 pub mod bloom;
 mod hash_queue;
+pub mod loader_utils;
 mod runtime;
 mod status_cache;
